+++ clients/algoliasearch-client-javascript/packages/ingestion/model/clientMethodProps.ts
@@ -1,13 +1,23 @@
 // This file is generated, manual changes will be lost - read more on https://github.com/algolia/api-clients-automation.
 
 import type { ActionType } from './actionType';
-import type { AuthenticationTypeWithNone } from './authenticationTypeWithNone';
+import type { AuthenticationSortKeys } from './authenticationSortKeys';
+import type { AuthenticationType } from './authenticationType';
 import type { AuthenticationUpdate } from './authenticationUpdate';
+import type { DestinationSortKeys } from './destinationSortKeys';
 import type { DestinationType } from './destinationType';
 import type { DestinationUpdate } from './destinationUpdate';
-import type { Platform } from './platform';
+import type { EventSortKeys } from './eventSortKeys';
+import type { EventStatus } from './eventStatus';
+import type { EventType } from './eventType';
+import type { OrderKeys } from './orderKeys';
+import type { PlatformWithNone } from './platformWithNone';
+import type { RunSortKeys } from './runSortKeys';
+import type { RunStatus } from './runStatus';
+import type { SourceSortKeys } from './sourceSortKeys';
 import type { SourceType } from './sourceType';
 import type { SourceUpdate } from './sourceUpdate';
+import type { TaskSortKeys } from './taskSortKeys';
 import type { TaskUpdate } from './taskUpdate';
 import type { TriggerType } from './triggerType';
 
@@ -16,7 +26,7 @@ import type { TriggerType } from './triggerType';
  */
 export type DeleteAuthenticationProps = {
   /**
-   * The authentication uuid.
+   * The authentication UUID.
    */
   authenticationID: string;
 };
@@ -26,7 +36,7 @@ export type DeleteAuthenticationProps = {
  */
 export type DeleteDestinationProps = {
   /**
-   * The destination uuid.
+   * The destination UUID.
    */
   destinationID: string;
 };
@@ -36,7 +46,7 @@ export type DeleteDestinationProps = {
  */
 export type DeleteSourceProps = {
   /**
-   * The source uuid.
+   * The source UUID.
    */
   sourceID: string;
 };
@@ -46,7 +56,7 @@ export type DeleteSourceProps = {
  */
 export type DeleteTaskProps = {
   /**
-   * The task uuid.
+   * The task UUID.
    */
   taskID: string;
 };
@@ -56,7 +66,7 @@ export type DeleteTaskProps = {
  */
 export type DisableTaskProps = {
   /**
-   * The task uuid.
+   * The task UUID.
    */
   taskID: string;
 };
@@ -66,7 +76,7 @@ export type DisableTaskProps = {
  */
 export type EnableTaskProps = {
   /**
-   * The task uuid.
+   * The task UUID.
    */
   taskID: string;
 };
@@ -76,7 +86,7 @@ export type EnableTaskProps = {
  */
 export type GetAuthenticationProps = {
   /**
-   * The authentication uuid.
+   * The authentication UUID.
    */
   authenticationID: string;
 };
@@ -94,13 +104,21 @@ export type GetAuthenticationsProps = {
    */
   page?: number;
   /**
-   * Which type the returned authentication should have. Can be a list of string separated with commas.
+   * The type of the authentications to retrieve.
    */
-  type?: AuthenticationTypeWithNone[];
+  type?: AuthenticationType[];
   /**
-   * Which platform the returned authentication should have. Can be a list of string separated with commas.
+   * The platform of the authentications to retrieve.
    */
-  platform?: Platform[];
+  platform?: PlatformWithNone[];
+  /**
+   * The key by which the list should be sorted.
+   */
+  sort?: AuthenticationSortKeys;
+  /**
+   * The order of the returned list.
+   */
+  order?: OrderKeys;
 };
 
 /**
@@ -108,7 +126,7 @@ export type GetAuthenticationsProps = {
  */
 export type GetDestinationProps = {
   /**
-   * The destination uuid.
+   * The destination UUID.
    */
   destinationID: string;
 };
@@ -126,13 +144,21 @@ export type GetDestinationsProps = {
    */
   page?: number;
   /**
-   * Which type the returned destination should have. Can be a list of string separated with commas.
+   * The type of the destinations to retrive.
    */
   type?: DestinationType[];
   /**
-   * Which authenticationID the returned destination should have. Can be a list of string separated with commas.
+   * The authenticationIDs of the destinations to retrive.
    */
   authenticationID?: string[];
+  /**
+   * The key by which the list should be sorted.
+   */
+  sort?: DestinationSortKeys;
+  /**
+   * The order of the returned list.
+   */
+  order?: OrderKeys;
 };
 
 /**
@@ -140,11 +166,11 @@ export type GetDestinationsProps = {
  */
 export type GetEventProps = {
   /**
-   * The run uuid.
+   * The run UUID.
    */
   runID: string;
   /**
-   * The event uuid.
+   * The event UUID.
    */
   eventID: string;
 };
@@ -154,17 +180,25 @@ export type GetEventProps = {
  */
 export type GetEventsProps = {
   /**
-   * The run uuid.
+   * The run UUID.
    */
   runID: string;
   /**
-   * The number of items per page to return.
+   * Filter the status of the events.
    */
-  itemsPerPage?: number;
+  status?: EventStatus[];
   /**
-   * The page number to fetch, starting at 1.
+   * Filter the type of the events.
    */
-  page?: number;
+  type?: EventType[];
+  /**
+   * The key by which the list should be sorted.
+   */
+  sort?: EventSortKeys;
+  /**
+   * The order of the returned list.
+   */
+  order?: OrderKeys;
 };
 
 /**
@@ -172,7 +206,7 @@ export type GetEventsProps = {
  */
 export type GetRunProps = {
   /**
-   * The run uuid.
+   * The run UUID.
    */
   runID: string;
 };
@@ -189,24 +223,22 @@ export type GetRunsProps = {
    * The page number to fetch, starting at 1.
    */
   page?: number;
-};
-
-/**
- * Properties for the `getRunsByTaskID` method.
- */
-export type GetRunsByTaskIDProps = {
   /**
-   * The task uuid.
+   * Filter the status of the runs.
    */
-  taskID: string;
+  status?: RunStatus[];
   /**
-   * The number of items per page to return.
+   * Filter by taskID.
    */
-  itemsPerPage?: number;
+  taskID?: string;
   /**
-   * The page number to fetch, starting at 1.
+   * The key by which the list should be sorted.
    */
-  page?: number;
+  sort?: RunSortKeys;
+  /**
+   * The order of the returned list.
+   */
+  order?: OrderKeys;
 };
 
 /**
@@ -214,7 +246,7 @@ export type GetRunsByTaskIDProps = {
  */
 export type GetSourceProps = {
   /**
-   * The source uuid.
+   * The source UUID.
    */
   sourceID: string;
 };
@@ -232,13 +264,21 @@ export type GetSourcesProps = {
    */
   page?: number;
   /**
-   * Which type the returned source should have. Can be a list of string separated with commas.
+   * The type of the sources to retrieve.
    */
   type?: SourceType[];
   /**
-   * Which authenticationID the returned source should have. Can be a list of string separated with commas.
+   * The authenticationIDs of the sources to retrieve. \'none\' returns sources that doesn\'t have an authentication.
    */
   authenticationID?: string[];
+  /**
+   * The key by which the list should be sorted.
+   */
+  sort?: SourceSortKeys;
+  /**
+   * The order of the returned list.
+   */
+  order?: OrderKeys;
 };
 
 /**
@@ -246,7 +286,7 @@ export type GetSourcesProps = {
  */
 export type GetTaskProps = {
   /**
-   * The task uuid.
+   * The task UUID.
    */
   taskID: string;
 };
@@ -264,21 +304,33 @@ export type GetTasksProps = {
    */
   page?: number;
   /**
-   * Which action the returned task should have. Can be a list of string separated with commas.
+   * The action of the tasks to retrieve.
    */
   action?: ActionType[];
   /**
-   * If the returned task should have its \'enabled\' property set to true.
+   * Whether the task is enabled or not.
    */
   enabled?: boolean;
   /**
-   * Which destinationID the returned task should have. Can be a list of string separated with commas.
+   * The sourceIDs of the tasks to retrive.
+   */
+  sourceID?: string[];
+  /**
+   * The destinationIDs of the tasks to retrive.
    */
   destinationID?: string[];
   /**
-   * Which trigger type the returned task should have. Can be a list of string separated with commas.
+   * The trigger type of the task.
    */
   triggerType?: TriggerType[];
+  /**
+   * The key by which the list should be sorted.
+   */
+  sort?: TaskSortKeys;
+  /**
+   * The order of the returned list.
+   */
+  order?: OrderKeys;
 };
 
 /**
@@ -286,7 +338,7 @@ export type GetTasksProps = {
  */
 export type RunTaskProps = {
   /**
-   * The task uuid.
+   * The task UUID.
    */
   taskID: string;
 };
@@ -296,7 +348,7 @@ export type RunTaskProps = {
  */
 export type UpdateAuthenticationProps = {
   /**
-   * The authentication uuid.
+   * The authentication UUID.
    */
   authenticationID: string;
   authenticationUpdate: AuthenticationUpdate;
@@ -307,7 +359,7 @@ export type UpdateAuthenticationProps = {
  */
 export type UpdateDestinationProps = {
   /**
-   * The destination uuid.
+   * The destination UUID.
    */
   destinationID: string;
   destinationUpdate: DestinationUpdate;
@@ -318,7 +370,7 @@ export type UpdateDestinationProps = {
  */
 export type UpdateSourceProps = {
   /**
-   * The source uuid.
+   * The source UUID.
    */
   sourceID: string;
   sourceUpdate: SourceUpdate;
@@ -329,7 +381,7 @@ export type UpdateSourceProps = {
  */
 export type UpdateTaskProps = {
   /**
-   * The task uuid.
+   * The task UUID.
    */
   taskID: string;
   taskUpdate: TaskUpdate;