+++ clients/algoliasearch-client-javascript/packages/ingestion/model/scheduleTriggerType.ts
@@ -1,3 +1,6 @@
 // This file is generated, manual changes will be lost - read more on https://github.com/algolia/api-clients-automation.
 
-export type AuthenticationNone = 'none';
+/**
+ * A task which is triggered by a schedule (cron expression).
+ */
+export type ScheduleTriggerType = 'schedule';