+++ clients/algoliasearch-client-javascript/packages/ingestion/model/sourceUpdateInput.ts
@@ -1,0 +1,10 @@
+// This file is generated, manual changes will be lost - read more on https://github.com/algolia/api-clients-automation.
+
+import type { SourceCSV } from './sourceCSV';
+import type { SourceJSON } from './sourceJSON';
+import type { SourceUpdateCommercetools } from './sourceUpdateCommercetools';
+
+export type SourceUpdateInput =
+  | SourceCSV
+  | SourceJSON
+  | SourceUpdateCommercetools;