+++ clients/algoliasearch-client-javascript/packages/ingestion/model/trigger.ts
@@ -1,19 +1,7 @@
 // This file is generated, manual changes will be lost - read more on https://github.com/algolia/api-clients-automation.
 
-import type { TriggerType } from './triggerType';
+import type { OnDemandTrigger } from './onDemandTrigger';
+import type { ScheduleTrigger } from './scheduleTrigger';
+import type { SubscriptionTrigger } from './subscriptionTrigger';
 
-export type Trigger = {
-  type: TriggerType;
-
-  cron?: string;
-
-  /**
-   * The last time the task ran (`scheduled` or `on-demand`).
-   */
-  lastRun?: string;
-
-  /**
-   * The next scheduled run for the task (`scheduled`).
-   */
-  nextRun?: string;
-};
+export type Trigger = OnDemandTrigger | ScheduleTrigger | SubscriptionTrigger;