+++ clients/algoliasearch-client-javascript/packages/ingestion/src/ingestionClient.ts
@@ -14,8 +14,10 @@ import type {
   QueryParameters,
 } from '@algolia/client-common';
 
+import type { Authentication } from '../model/authentication';
 import type { AuthenticationCreate } from '../model/authenticationCreate';
 import type { AuthenticationCreateResponse } from '../model/authenticationCreateResponse';
+import type { AuthenticationSearch } from '../model/authenticationSearch';
 import type { AuthenticationUpdateResponse } from '../model/authenticationUpdateResponse';
 import type { AuthenticationWithInput } from '../model/authenticationWithInput';
 import type {
@@ -33,7 +35,6 @@ import type {
   GetEventsProps,
   GetRunProps,
   GetRunsProps,
-  GetRunsByTaskIDProps,
   GetSourceProps,
   GetSourcesProps,
   GetTaskProps,
@@ -48,6 +49,7 @@ import type { DeleteResponse } from '../model/deleteResponse';
 import type { Destination } from '../model/destination';
 import type { DestinationCreate } from '../model/destinationCreate';
 import type { DestinationCreateResponse } from '../model/destinationCreateResponse';
+import type { DestinationSearch } from '../model/destinationSearch';
 import type { DestinationUpdateResponse } from '../model/destinationUpdateResponse';
 import type { Event } from '../model/event';
 import type { ListAuthenticationsResponse } from '../model/listAuthenticationsResponse';
@@ -55,19 +57,25 @@ import type { ListDestinationsResponse } from '../model/listDestinationsResponse
 import type { ListEventsResponse } from '../model/listEventsResponse';
 import type { ListSourcesResponse } from '../model/listSourcesResponse';
 import type { ListTasksResponse } from '../model/listTasksResponse';
+import type { OnDemandTrigger } from '../model/onDemandTrigger';
 import type { Run } from '../model/run';
 import type { RunListResponse } from '../model/runListResponse';
 import type { RunResponse } from '../model/runResponse';
+import type { ScheduleTrigger } from '../model/scheduleTrigger';
 import type { Source } from '../model/source';
 import type { SourceCreate } from '../model/sourceCreate';
 import type { SourceCreateResponse } from '../model/sourceCreateResponse';
+import type { SourceSearch } from '../model/sourceSearch';
 import type { SourceUpdateResponse } from '../model/sourceUpdateResponse';
+import type { SubscriptionTrigger } from '../model/subscriptionTrigger';
 import type { Task } from '../model/task';
 import type { TaskCreate } from '../model/taskCreate';
 import type { TaskCreateResponse } from '../model/taskCreateResponse';
+import type { TaskSearch } from '../model/taskSearch';
 import type { TaskUpdateResponse } from '../model/taskUpdateResponse';
+import type { Trigger } from '../model/trigger';
 
-export const apiClientVersion = '1.0.0-alpha.10';
+export const apiClientVersion = '1.0.0-alpha.18';
 
 export const REGIONS = ['us'] as const;
 export type Region = typeof REGIONS[number];
@@ -76,6 +84,42 @@ function getDefaultHosts(region: Region): Host[] {
   const url = 'data.{region}.algolia.com'.replace('{region}', region);
 
   return [{ url, accept: 'readWrite', protocol: 'https' }];
+}
+
+/**
+ * Guard: Return strongly typed specific OnDemandTrigger for a given Trigger.
+ *
+ * @summary Guard method that returns a strongly typed specific OnDemandTrigger for a given Trigger.
+ * @param trigger - The given Task Trigger.
+ */
+export function isOnDemandTrigger(
+  trigger: Trigger
+): trigger is OnDemandTrigger {
+  return trigger.type === 'onDemand';
+}
+
+/**
+ * Guard: Return strongly typed specific ScheduleTrigger for a given Trigger.
+ *
+ * @summary Guard method that returns a strongly typed specific ScheduleTrigger for a given Trigger.
+ * @param trigger - The given Task Trigger.
+ */
+export function isScheduleTrigger(
+  trigger: Trigger
+): trigger is ScheduleTrigger {
+  return trigger.type === 'schedule';
+}
+
+/**
+ * Guard: Return strongly typed specific SubscriptionTrigger for a given Trigger.
+ *
+ * @summary Guard method that returns a strongly typed specific SubscriptionTrigger for a given Trigger.
+ * @param trigger - The given Task Trigger.
+ */
+export function isSubscriptionTrigger(
+  trigger: Trigger
+): trigger is SubscriptionTrigger {
+  return trigger.type === 'subscription';
 }
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
@@ -226,6 +270,17 @@ export function createIngestionClient({
         );
       }
 
+      if (!sourceCreate.type) {
+        throw new Error(
+          'Parameter `sourceCreate.type` is required when calling `createSource`.'
+        );
+      }
+      if (!sourceCreate.input) {
+        throw new Error(
+          'Parameter `sourceCreate.input` is required when calling `createSource`.'
+        );
+      }
+
       const requestPath = '/1/sources';
       const headers: Headers = {};
       const queryParameters: QueryParameters = {};
@@ -263,6 +318,11 @@ export function createIngestionClient({
           'Parameter `taskCreate.sourceID` is required when calling `createTask`.'
         );
       }
+      if (!taskCreate.trigger) {
+        throw new Error(
+          'Parameter `taskCreate.trigger` is required when calling `createTask`.'
+        );
+      }
       if (!taskCreate.action) {
         throw new Error(
           'Parameter `taskCreate.action` is required when calling `createTask`.'
@@ -285,11 +345,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Soft delete a authentication.
+     * Soft delete the authentication of the given authenticationID.
      *
      * @summary Delete a authentication.
      * @param deleteAuthentication - The deleteAuthentication object.
-     * @param deleteAuthentication.authenticationID - The authentication uuid.
+     * @param deleteAuthentication.authenticationID - The authentication UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     deleteAuthentication(
@@ -320,11 +380,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Soft delete a destination.
+     * Soft delete the destination of the given destinationID.
      *
      * @summary Delete a destination.
      * @param deleteDestination - The deleteDestination object.
-     * @param deleteDestination.destinationID - The destination uuid.
+     * @param deleteDestination.destinationID - The destination UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     deleteDestination(
@@ -355,11 +415,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Soft delete a source.
+     * Soft delete the source of the given sourceID.
      *
      * @summary Delete a source.
      * @param deleteSource - The deleteSource object.
-     * @param deleteSource.sourceID - The source uuid.
+     * @param deleteSource.sourceID - The source UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     deleteSource(
@@ -390,11 +450,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Soft delete a task.
+     * Soft delete the task of the given taskID.
      *
      * @summary Delete a task.
      * @param deleteTask - The deleteTask object.
-     * @param deleteTask.taskID - The task uuid.
+     * @param deleteTask.taskID - The task UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     deleteTask(
@@ -425,11 +485,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Disable a task.
+     * Disable the task of the given taskID.
      *
      * @summary Disable a task.
      * @param disableTask - The disableTask object.
-     * @param disableTask.taskID - The task uuid.
+     * @param disableTask.taskID - The task UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     disableTask(
@@ -460,11 +520,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Enable a task.
+     * Enable the task of the given taskID.
      *
      * @summary Enable a task.
      * @param enableTask - The enableTask object.
-     * @param enableTask.taskID - The task uuid.
+     * @param enableTask.taskID - The task UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     enableTask(
@@ -495,11 +555,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a authentication.
+     * Get the authentication of the given authenticationID.
      *
      * @summary Get a authentication.
      * @param getAuthentication - The getAuthentication object.
-     * @param getAuthentication.authenticationID - The authentication uuid.
+     * @param getAuthentication.authenticationID - The authentication UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getAuthentication(
@@ -530,18 +590,27 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a list of authentications.
+     * Get a list of authentications for the given query parameters, with pagination details.
      *
      * @summary Get a list of authentications.
      * @param getAuthentications - The getAuthentications object.
      * @param getAuthentications.itemsPerPage - The number of items per page to return.
      * @param getAuthentications.page - The page number to fetch, starting at 1.
-     * @param getAuthentications.type - Which type the returned authentication should have. Can be a list of string separated with commas.
-     * @param getAuthentications.platform - Which platform the returned authentication should have. Can be a list of string separated with commas.
+     * @param getAuthentications.type - The type of the authentications to retrieve.
+     * @param getAuthentications.platform - The platform of the authentications to retrieve.
+     * @param getAuthentications.sort - The key by which the list should be sorted.
+     * @param getAuthentications.order - The order of the returned list.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getAuthentications(
-      { itemsPerPage, page, type, platform }: GetAuthenticationsProps = {},
+      {
+        itemsPerPage,
+        page,
+        type,
+        platform,
+        sort,
+        order,
+      }: GetAuthenticationsProps = {},
       requestOptions: RequestOptions | undefined = undefined
     ): Promise<ListAuthenticationsResponse> {
       const requestPath = '/1/authentications';
@@ -564,6 +633,14 @@ export function createIngestionClient({
         queryParameters.platform = platform.toString();
       }
 
+      if (sort !== undefined) {
+        queryParameters.sort = sort.toString();
+      }
+
+      if (order !== undefined) {
+        queryParameters.order = order.toString();
+      }
+
       const request: Request = {
         method: 'GET',
         path: requestPath,
@@ -575,11 +652,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a destination.
+     * Get the destination of the given destinationID.
      *
      * @summary Get a destination.
      * @param getDestination - The getDestination object.
-     * @param getDestination.destinationID - The destination uuid.
+     * @param getDestination.destinationID - The destination UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getDestination(
@@ -610,18 +687,27 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a list of destinations.
+     * Get a list of destinations for the given query parameters, with pagination details.
      *
      * @summary Get a list of destinations.
      * @param getDestinations - The getDestinations object.
      * @param getDestinations.itemsPerPage - The number of items per page to return.
      * @param getDestinations.page - The page number to fetch, starting at 1.
-     * @param getDestinations.type - Which type the returned destination should have. Can be a list of string separated with commas.
-     * @param getDestinations.authenticationID - Which authenticationID the returned destination should have. Can be a list of string separated with commas.
+     * @param getDestinations.type - The type of the destinations to retrive.
+     * @param getDestinations.authenticationID - The authenticationIDs of the destinations to retrive.
+     * @param getDestinations.sort - The key by which the list should be sorted.
+     * @param getDestinations.order - The order of the returned list.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getDestinations(
-      { itemsPerPage, page, type, authenticationID }: GetDestinationsProps = {},
+      {
+        itemsPerPage,
+        page,
+        type,
+        authenticationID,
+        sort,
+        order,
+      }: GetDestinationsProps = {},
       requestOptions: RequestOptions | undefined = undefined
     ): Promise<ListDestinationsResponse> {
       const requestPath = '/1/destinations';
@@ -644,6 +730,14 @@ export function createIngestionClient({
         queryParameters.authenticationID = authenticationID.toString();
       }
 
+      if (sort !== undefined) {
+        queryParameters.sort = sort.toString();
+      }
+
+      if (order !== undefined) {
+        queryParameters.order = order.toString();
+      }
+
       const request: Request = {
         method: 'GET',
         path: requestPath,
@@ -655,12 +749,12 @@ export function createIngestionClient({
     },
 
     /**
-     * Get an event in a specific run.
+     * Get a single event for a specific runID.
      *
-     * @summary Get an event in a specific run.
+     * @summary Get an event.
      * @param getEvent - The getEvent object.
-     * @param getEvent.runID - The run uuid.
-     * @param getEvent.eventID - The event uuid.
+     * @param getEvent.runID - The run UUID.
+     * @param getEvent.eventID - The event UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getEvent(
@@ -696,17 +790,19 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a list of events for a specific run.
+     * Get a list of events associated to the given runID, for the given query parameters.
      *
-     * @summary Get a list of events for a specific run.
+     * @summary Get a list of events.
      * @param getEvents - The getEvents object.
-     * @param getEvents.runID - The run uuid.
-     * @param getEvents.itemsPerPage - The number of items per page to return.
-     * @param getEvents.page - The page number to fetch, starting at 1.
+     * @param getEvents.runID - The run UUID.
+     * @param getEvents.status - Filter the status of the events.
+     * @param getEvents.type - Filter the type of the events.
+     * @param getEvents.sort - The key by which the list should be sorted.
+     * @param getEvents.order - The order of the returned list.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getEvents(
-      { runID, itemsPerPage, page }: GetEventsProps,
+      { runID, status, type, sort, order }: GetEventsProps,
       requestOptions?: RequestOptions
     ): Promise<ListEventsResponse> {
       if (!runID) {
@@ -722,12 +818,20 @@ export function createIngestionClient({
       const headers: Headers = {};
       const queryParameters: QueryParameters = {};
 
-      if (itemsPerPage !== undefined) {
-        queryParameters.itemsPerPage = itemsPerPage.toString();
+      if (status !== undefined) {
+        queryParameters.status = status.toString();
       }
 
-      if (page !== undefined) {
-        queryParameters.page = page.toString();
+      if (type !== undefined) {
+        queryParameters.type = type.toString();
+      }
+
+      if (sort !== undefined) {
+        queryParameters.sort = sort.toString();
+      }
+
+      if (order !== undefined) {
+        queryParameters.order = order.toString();
       }
 
       const request: Request = {
@@ -741,11 +845,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a run.
+     * Get a single run for the given ID.
      *
      * @summary Get a run.
      * @param getRun - The getRun object.
-     * @param getRun.runID - The run uuid.
+     * @param getRun.runID - The run UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getRun(
@@ -774,16 +878,20 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a list of runs.
+     * Get a list of runs for the given query parameters, with pagination details.
      *
      * @summary Get a list of runs.
      * @param getRuns - The getRuns object.
      * @param getRuns.itemsPerPage - The number of items per page to return.
      * @param getRuns.page - The page number to fetch, starting at 1.
+     * @param getRuns.status - Filter the status of the runs.
+     * @param getRuns.taskID - Filter by taskID.
+     * @param getRuns.sort - The key by which the list should be sorted.
+     * @param getRuns.order - The order of the returned list.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getRuns(
-      { itemsPerPage, page }: GetRunsProps = {},
+      { itemsPerPage, page, status, taskID, sort, order }: GetRunsProps = {},
       requestOptions: RequestOptions | undefined = undefined
     ): Promise<RunListResponse> {
       const requestPath = '/1/runs';
@@ -798,49 +906,20 @@ export function createIngestionClient({
         queryParameters.page = page.toString();
       }
 
-      const request: Request = {
-        method: 'GET',
-        path: requestPath,
-        queryParameters,
-        headers,
-      };
-
-      return transporter.request(request, requestOptions);
-    },
-
-    /**
-     * Get a list of runs associated with a taskID.
-     *
-     * @summary Get a list of runs associated with a taskID.
-     * @param getRunsByTaskID - The getRunsByTaskID object.
-     * @param getRunsByTaskID.taskID - The task uuid.
-     * @param getRunsByTaskID.itemsPerPage - The number of items per page to return.
-     * @param getRunsByTaskID.page - The page number to fetch, starting at 1.
-     * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
-     */
-    getRunsByTaskID(
-      { taskID, itemsPerPage, page }: GetRunsByTaskIDProps,
-      requestOptions?: RequestOptions
-    ): Promise<RunListResponse> {
-      if (!taskID) {
-        throw new Error(
-          'Parameter `taskID` is required when calling `getRunsByTaskID`.'
-        );
+      if (status !== undefined) {
+        queryParameters.status = status.toString();
       }
 
-      const requestPath = '/1/runs/tasks/{taskID}'.replace(
-        '{taskID}',
-        encodeURIComponent(taskID)
-      );
-      const headers: Headers = {};
-      const queryParameters: QueryParameters = {};
-
-      if (itemsPerPage !== undefined) {
-        queryParameters.itemsPerPage = itemsPerPage.toString();
+      if (taskID !== undefined) {
+        queryParameters.taskID = taskID.toString();
       }
 
-      if (page !== undefined) {
-        queryParameters.page = page.toString();
+      if (sort !== undefined) {
+        queryParameters.sort = sort.toString();
+      }
+
+      if (order !== undefined) {
+        queryParameters.order = order.toString();
       }
 
       const request: Request = {
@@ -854,11 +933,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a source.
+     * Get the source of the given sourceID.
      *
      * @summary Get a source.
      * @param getSource - The getSource object.
-     * @param getSource.sourceID - The source uuid.
+     * @param getSource.sourceID - The source UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getSource(
@@ -889,18 +968,27 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a list of sources.
+     * Get a list of sources for the given query parameters, with pagination details.
      *
      * @summary Get a list of sources.
      * @param getSources - The getSources object.
      * @param getSources.itemsPerPage - The number of items per page to return.
      * @param getSources.page - The page number to fetch, starting at 1.
-     * @param getSources.type - Which type the returned source should have. Can be a list of string separated with commas.
-     * @param getSources.authenticationID - Which authenticationID the returned source should have. Can be a list of string separated with commas.
+     * @param getSources.type - The type of the sources to retrieve.
+     * @param getSources.authenticationID - The authenticationIDs of the sources to retrieve. \'none\' returns sources that doesn\'t have an authentication.
+     * @param getSources.sort - The key by which the list should be sorted.
+     * @param getSources.order - The order of the returned list.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getSources(
-      { itemsPerPage, page, type, authenticationID }: GetSourcesProps = {},
+      {
+        itemsPerPage,
+        page,
+        type,
+        authenticationID,
+        sort,
+        order,
+      }: GetSourcesProps = {},
       requestOptions: RequestOptions | undefined = undefined
     ): Promise<ListSourcesResponse> {
       const requestPath = '/1/sources';
@@ -923,6 +1011,14 @@ export function createIngestionClient({
         queryParameters.authenticationID = authenticationID.toString();
       }
 
+      if (sort !== undefined) {
+        queryParameters.sort = sort.toString();
+      }
+
+      if (order !== undefined) {
+        queryParameters.order = order.toString();
+      }
+
       const request: Request = {
         method: 'GET',
         path: requestPath,
@@ -934,11 +1030,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a task.
+     * Get the task of the given taskID.
      *
      * @summary Get a task.
      * @param getTask - The getTask object.
-     * @param getTask.taskID - The task uuid.
+     * @param getTask.taskID - The task UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getTask(
@@ -969,16 +1065,19 @@ export function createIngestionClient({
     },
 
     /**
-     * Get a list of tasks.
+     * Get a list of tasks for the given query parameters, with pagination details.
      *
      * @summary Get a list of tasks.
      * @param getTasks - The getTasks object.
      * @param getTasks.itemsPerPage - The number of items per page to return.
      * @param getTasks.page - The page number to fetch, starting at 1.
-     * @param getTasks.action - Which action the returned task should have. Can be a list of string separated with commas.
-     * @param getTasks.enabled - If the returned task should have its \'enabled\' property set to true.
-     * @param getTasks.destinationID - Which destinationID the returned task should have. Can be a list of string separated with commas.
-     * @param getTasks.triggerType - Which trigger type the returned task should have. Can be a list of string separated with commas.
+     * @param getTasks.action - The action of the tasks to retrieve.
+     * @param getTasks.enabled - Whether the task is enabled or not.
+     * @param getTasks.sourceID - The sourceIDs of the tasks to retrive.
+     * @param getTasks.destinationID - The destinationIDs of the tasks to retrive.
+     * @param getTasks.triggerType - The trigger type of the task.
+     * @param getTasks.sort - The key by which the list should be sorted.
+     * @param getTasks.order - The order of the returned list.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     getTasks(
@@ -987,8 +1086,11 @@ export function createIngestionClient({
         page,
         action,
         enabled,
+        sourceID,
         destinationID,
         triggerType,
+        sort,
+        order,
       }: GetTasksProps = {},
       requestOptions: RequestOptions | undefined = undefined
     ): Promise<ListTasksResponse> {
@@ -1012,12 +1114,24 @@ export function createIngestionClient({
         queryParameters.enabled = enabled.toString();
       }
 
+      if (sourceID !== undefined) {
+        queryParameters.sourceID = sourceID.toString();
+      }
+
       if (destinationID !== undefined) {
         queryParameters.destinationID = destinationID.toString();
       }
 
       if (triggerType !== undefined) {
         queryParameters.triggerType = triggerType.toString();
+      }
+
+      if (sort !== undefined) {
+        queryParameters.sort = sort.toString();
+      }
+
+      if (order !== undefined) {
+        queryParameters.order = order.toString();
       }
 
       const request: Request = {
@@ -1031,11 +1145,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Run a task.
+     * Run the task of the given taskID.
      *
      * @summary Run a task.
      * @param runTask - The runTask object.
-     * @param runTask.taskID - The task uuid.
+     * @param runTask.taskID - The task UUID.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
     runTask(
@@ -1066,11 +1180,163 @@ export function createIngestionClient({
     },
 
     /**
-     * Update a authentication.
+     * Search among authentications with a defined set of parameters.
+     *
+     * @summary Search among authentications.
+     * @param authenticationSearch - The authenticationSearch object.
+     * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
+     */
+    searchAuthentications(
+      authenticationSearch: AuthenticationSearch,
+      requestOptions?: RequestOptions
+    ): Promise<Authentication[]> {
+      if (!authenticationSearch) {
+        throw new Error(
+          'Parameter `authenticationSearch` is required when calling `searchAuthentications`.'
+        );
+      }
+
+      if (!authenticationSearch.authenticationIDs) {
+        throw new Error(
+          'Parameter `authenticationSearch.authenticationIDs` is required when calling `searchAuthentications`.'
+        );
+      }
+
+      const requestPath = '/1/authentications/search';
+      const headers: Headers = {};
+      const queryParameters: QueryParameters = {};
+
+      const request: Request = {
+        method: 'POST',
+        path: requestPath,
+        queryParameters,
+        headers,
+        data: authenticationSearch,
+      };
+
+      return transporter.request(request, requestOptions);
+    },
+
+    /**
+     * Search among destinations with a defined set of parameters.
+     *
+     * @summary Search among destinations.
+     * @param destinationSearch - The destinationSearch object.
+     * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
+     */
+    searchDestinations(
+      destinationSearch: DestinationSearch,
+      requestOptions?: RequestOptions
+    ): Promise<Destination[]> {
+      if (!destinationSearch) {
+        throw new Error(
+          'Parameter `destinationSearch` is required when calling `searchDestinations`.'
+        );
+      }
+
+      if (!destinationSearch.destinationIDs) {
+        throw new Error(
+          'Parameter `destinationSearch.destinationIDs` is required when calling `searchDestinations`.'
+        );
+      }
+
+      const requestPath = '/1/destinations/search';
+      const headers: Headers = {};
+      const queryParameters: QueryParameters = {};
+
+      const request: Request = {
+        method: 'POST',
+        path: requestPath,
+        queryParameters,
+        headers,
+        data: destinationSearch,
+      };
+
+      return transporter.request(request, requestOptions);
+    },
+
+    /**
+     * Search among sources with a defined set of parameters.
+     *
+     * @summary Search among sources.
+     * @param sourceSearch - The sourceSearch object.
+     * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
+     */
+    searchSources(
+      sourceSearch: SourceSearch,
+      requestOptions?: RequestOptions
+    ): Promise<Source[]> {
+      if (!sourceSearch) {
+        throw new Error(
+          'Parameter `sourceSearch` is required when calling `searchSources`.'
+        );
+      }
+
+      if (!sourceSearch.sourceIDs) {
+        throw new Error(
+          'Parameter `sourceSearch.sourceIDs` is required when calling `searchSources`.'
+        );
+      }
+
+      const requestPath = '/1/sources/search';
+      const headers: Headers = {};
+      const queryParameters: QueryParameters = {};
+
+      const request: Request = {
+        method: 'POST',
+        path: requestPath,
+        queryParameters,
+        headers,
+        data: sourceSearch,
+      };
+
+      return transporter.request(request, requestOptions);
+    },
+
+    /**
+     * Search among tasks with a defined set of parameters.
+     *
+     * @summary Search among tasks.
+     * @param taskSearch - The taskSearch object.
+     * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
+     */
+    searchTasks(
+      taskSearch: TaskSearch,
+      requestOptions?: RequestOptions
+    ): Promise<Task[]> {
+      if (!taskSearch) {
+        throw new Error(
+          'Parameter `taskSearch` is required when calling `searchTasks`.'
+        );
+      }
+
+      if (!taskSearch.taskIDs) {
+        throw new Error(
+          'Parameter `taskSearch.taskIDs` is required when calling `searchTasks`.'
+        );
+      }
+
+      const requestPath = '/1/tasks/search';
+      const headers: Headers = {};
+      const queryParameters: QueryParameters = {};
+
+      const request: Request = {
+        method: 'POST',
+        path: requestPath,
+        queryParameters,
+        headers,
+        data: taskSearch,
+      };
+
+      return transporter.request(request, requestOptions);
+    },
+
+    /**
+     * Update the authentication of the given authenticationID.
      *
      * @summary Update a authentication.
      * @param updateAuthentication - The updateAuthentication object.
-     * @param updateAuthentication.authenticationID - The authentication uuid.
+     * @param updateAuthentication.authenticationID - The authentication UUID.
      * @param updateAuthentication.authenticationUpdate - The authenticationUpdate object.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
@@ -1109,11 +1375,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Update a destination.
+     * Update the destination of the given destinationID.
      *
      * @summary Update a destination.
      * @param updateDestination - The updateDestination object.
-     * @param updateDestination.destinationID - The destination uuid.
+     * @param updateDestination.destinationID - The destination UUID.
      * @param updateDestination.destinationUpdate - The destinationUpdate object.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
@@ -1152,11 +1418,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Update a source.
+     * Update the source of the given sourceID.
      *
      * @summary Update a source.
      * @param updateSource - The updateSource object.
-     * @param updateSource.sourceID - The source uuid.
+     * @param updateSource.sourceID - The source UUID.
      * @param updateSource.sourceUpdate - The sourceUpdate object.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */
@@ -1184,7 +1450,7 @@ export function createIngestionClient({
       const queryParameters: QueryParameters = {};
 
       const request: Request = {
-        method: 'POST',
+        method: 'PUT',
         path: requestPath,
         queryParameters,
         headers,
@@ -1195,11 +1461,11 @@ export function createIngestionClient({
     },
 
     /**
-     * Update a task.
+     * Update the task of the given taskID.
      *
      * @summary Update a task.
      * @param updateTask - The updateTask object.
-     * @param updateTask.taskID - The task uuid.
+     * @param updateTask.taskID - The task UUID.
      * @param updateTask.taskUpdate - The taskUpdate object.
      * @param requestOptions - The requestOptions to send along with the query, they will be merged with the transporter requestOptions.
      */