+++ clients/algoliasearch-client-php/CHANGELOG.md
@@ -1,3 +1,34 @@
+## [4.0.0-alpha.42](https://github.com/algolia/algoliasearch-client-php/compare/4.0.0-alpha.41...4.0.0-alpha.42)
+
+- [023d5572](https://github.com/algolia/api-clients-automation/commit/023d5572) feat(specs): add sort and order handling in query parameters ([#1304](https://github.com/algolia/api-clients-automation/pull/1304)) by [@Fluf22](https://github.com/Fluf22/)
+
+## [4.0.0-alpha.41](https://github.com/algolia/algoliasearch-client-php/compare/4.0.0-alpha.40...4.0.0-alpha.41)
+
+- [b9b4ba35](https://github.com/algolia/api-clients-automation/commit/b9b4ba35) feat(specs): add CSV source ([#1301](https://github.com/algolia/api-clients-automation/pull/1301)) by [@morganleroi](https://github.com/morganleroi/)
+
+## [4.0.0-alpha.40](https://github.com/algolia/algoliasearch-client-php/compare/4.0.0-alpha.39...4.0.0-alpha.40)
+
+- [9d04dc22](https://github.com/algolia/api-clients-automation/commit/9d04dc22) fix(specs): rename task type to `onDemand` ([#1299](https://github.com/algolia/api-clients-automation/pull/1299)) by [@shortcuts](https://github.com/shortcuts/)
+
+## [4.0.0-alpha.39](https://github.com/algolia/algoliasearch-client-php/compare/4.0.0-alpha.38...4.0.0-alpha.39)
+
+- [a1c7c245](https://github.com/algolia/api-clients-automation/commit/a1c7c245) fix(specs): correct types for Tasks trigger ([#1279](https://github.com/algolia/api-clients-automation/pull/1279)) by [@shortcuts](https://github.com/shortcuts/)
+
+## [4.0.0-alpha.38](https://github.com/algolia/algoliasearch-client-php/compare/4.0.0-alpha.37...4.0.0-alpha.38)
+
+- [aed20fe6](https://github.com/algolia/api-clients-automation/commit/aed20fe6) feat(specs): add more descriptions to the Ingestion API ([#1273](https://github.com/algolia/api-clients-automation/pull/1273)) by [@shortcuts](https://github.com/shortcuts/)
+- [3739451a](https://github.com/algolia/api-clients-automation/commit/3739451a) fix(specs): add query params for observability ([#1277](https://github.com/algolia/api-clients-automation/pull/1277)) by [@millotp](https://github.com/millotp/)
+- [859c0a97](https://github.com/algolia/api-clients-automation/commit/859c0a97) feat(specs): source update changes ([#1276](https://github.com/algolia/api-clients-automation/pull/1276)) by [@mehmetaligok](https://github.com/mehmetaligok/)
+
+## [4.0.0-alpha.37](https://github.com/algolia/algoliasearch-client-php/compare/4.0.0-alpha.36...4.0.0-alpha.37)
+
+- [46c1e2e9](https://github.com/algolia/api-clients-automation/commit/46c1e2e9) feat(specs): add search endpoint to every item type ([#1274](https://github.com/algolia/api-clients-automation/pull/1274)) by [@Fluf22](https://github.com/Fluf22/)
+- [736bc4aa](https://github.com/algolia/api-clients-automation/commit/736bc4aa) fix(php): fix deserialization of models ([#1265](https://github.com/algolia/api-clients-automation/pull/1265)) by [@damcou](https://github.com/damcou/)
+
+## [4.0.0-alpha.36](https://github.com/algolia/algoliasearch-client-php/compare/4.0.0-alpha.35...4.0.0-alpha.36)
+
+- [a60e6790](https://github.com/algolia/api-clients-automation/commit/a60e6790) feat(specs): add sourceID query param to task list endpoint ([#1250](https://github.com/algolia/api-clients-automation/pull/1250)) by [@Fluf22](https://github.com/Fluf22/)
+
 ## [4.0.0-alpha.35](https://github.com/algolia/algoliasearch-client-php/compare/4.0.0-alpha.34...4.0.0-alpha.35)
 
 - [83330afe](https://github.com/algolia/api-clients-automation/commit/83330afe) chore(specs): change PlatformType naming to Platform ([#1246](https://github.com/algolia/api-clients-automation/pull/1246)) by [@Fluf22](https://github.com/Fluf22/)